Maintainer changed | Description area specified
You can add appearance comments here to be presented in the high-tech editors such as PyCharm, VsCode, and etc. 
The contents may contain a bunch of information such as descryptions, usages, authors, and whatever you think would be
nice to be presented. 

take a look at this file.
https://github.com/lnxpy/PyTeX/blob/master/pytex.py

Try to clone the PyTeX and hover your mouse on the imported module. You will see the decryption as well.